feat: parallelize summaries
