+++ src/s3_log_extraction/extractors/_s3_log_access_extractor.py
@@ -13,7 +13,8 @@ import natsort
 import tqdm
 
 from ._globals import _STOP_EXTRACTION_FILE_NAME
-from ._utils import _deploy_subprocess, _handle_gawk_base, _handle_max_workers
+from ._utils import _deploy_subprocess, _handle_gawk_base
+from .._parallel._utils import _handle_max_workers
 from ..config import get_cache_directory, get_extraction_directory, get_records_directory
